Add Carbon ad after export analysis
